create a BaseSQLiteApiExt for DllImported platforms
