+++ src/SQLite.Net.Platform.Generic/SQLite.Net.Platform.Generic.csproj
@@ -52,4 +52,5 @@
     </ProjectReference>
   </ItemGroup>
   <Import Project="..\SQLite.Net.Platform.Shared.Interop\SQLite.Net.Platform.Shared.Interop.projitems" Label="Shared" />
+  <Import Project="..\SQLite.Net.Platform.Shared.Common\SQLite.Net.Platform.Shared.Common.projitems" Label="Shared" />
 </Project>
