+++ src/SQLite.Net.Platform.Win32/SQLite.Net.Platform.Win32.csproj
@@ -73,6 +73,7 @@
     <None Include="packages.config" />
   </ItemGroup>
   <Import Project="..\SQLite.Net.Platform.Shared.Interop\SQLite.Net.Platform.Shared.Interop.projitems" Label="Shared" />
+  <Import Project="..\SQLite.Net.Platform.Shared.Common\SQLite.Net.Platform.Shared.Common.projitems" Label="Shared" />
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
   <Import Project="$(SolutionDir)\.nuget\NuGet.targets" Condition="Exists('$(SolutionDir)\.nuget\NuGet.targets')" />
   <!-- To modify your build process, add your task inside one of the targets below and uncomment it.
